refactor(engine,state,db): Seperate engine and state pkgs, move db code into state
The engine package drives the data represented in state by querying for
Eblocks and passing them to state.ProcessEBlock().

The state package represents the state of various chains. It maintains a
global RWMutex protected chainMap that indexes the chain struct by its
chain ID. It maintains the open db connection to that chain's db and a
cache of some chain data.

Current database access is tightly coupled with the state. In the
future, database related functions may be moved to their own db package.
But that proved overly complex in the initial design.