feat(fat1,factom): Add support for fat1.NFTokens
Introduce the basic data types that form FAT-1 transactions.
AddressNFTokensMap is a type that maps RCDHashes to NFTokens. NFTokens
represents a unique set of NFTokenIDs. NFTokenIDs are integers and may
be represented individually or as an NFTokenIDRange.

AddressNFTokensMap unmarshals a JSON object with FA addresses as field
keys and arrays of NFTokenIDs as the values. The unmarshaling strictly
enforces that the JSON object contain no duplicate key values, and that
there are no duplicate NFTokenIDs within or between FA addresses.

- Improve factom.Address by adding RCDHash and PrivateKey types
Since we work with RCD hashes directly as addresses it made sense to
have a dedicated type that implements unmarshaling and marshaling of
human readable factoid addresses. Really an RCDHash could hold the hash
of any RCD and the appropriate human readable representation of it may
not necessarily be an factoid address, but that's the majority of the
use case for our purposes at the moment. This also introduces a
factom.PrivateKey type that unmarshals Fs secret keys. If you attempt to
unmarshal either an FA or Fs JSON string into an Address, the correct
type will be used.