feat(factom): Add type NetworkID, Bytes32{}.IsZero(), other refactors
The type NetworkID provides convenience features for working with the 4
magic bytes that identify different factom networks. MainnetID() and
TestnetID() return the official mainnet and testnet Network IDs.

Bytes32{}.IsZero() is a convenience function for checking if a Bytes32
object is zero.

EBlock{}.GetAllPrev() was refactored to be easier to understand and
correct. The behavior was also changed: the order of the returned
[]EBlock is now opposite, with the last element being the first EBlock
in the chain.

The Height field in Entry{} has been removed as it is not actually part
of the binary datastructure and was only populated if the Entry{} was
created by a call to EBlock{}.Get(). Now the Height needs to be passed
separately. Consequently, Identity{} now has a Height field since it is
no longer part of the embedded Entry{}.

Identity{}.UnmarshalBinary unmarshals Entry data and verifies internal
consistency before populating Identity{}.